Moved moved matrix.py to mobject, and pulled out MatrixMultiplicationScene
